backend/requirements.txt
fastapi[standard]
uvicorn
pandas
openpyxl
python-multipart
google-genai
python-dotenv
